# Use RECV syscall instead of READ in do_read_some()