`recv` has a slight performance advantage over the more generic `read`. Seastar is already using `send` instead of `write` on the transmit side. 

N.B. I had to change the return type from `ssize_t` to `size_t` in posix.hh to get it working.