@talawahtech 2022-06-23
Sure, I did my tests using a simple (fake) HTTP server. It sends the same HTTP response to any incoming TCP request. The server ran on a  4 vCPU c5n.xlarge AWS instance, while the client ran wrk (slightly modified) on a 16 vCPU c5n.4xlarge instance. Performance improved from 964k req/s to 982k req/s, a 1.9% performance improvement.

Before
```
Running 5s test @ http://172.31.XX.XX:8080/json
  16 threads and 256 connections
  Latency Distribution
  50.00%  257.00us
  90.00%  296.00us
  99.00%  337.00us
  4820680 requests in 5.00s, 712.59MB read
Requests/sec: 964121.54
```

After
```
Running 5s test @ http://172.31.XX.XX:8080/json
  16 threads and 256 connections
  Latency Distribution
  50.00%  250.00us
  90.00%  291.00us
  99.00%  342.00us
  4911503 requests in 5.00s, 726.02MB read
Requests/sec: 982287.44
```