@psarna 2022-06-24
Thanks! @avikivity, do you consider the ssize_t->size_t migration a breaking change or can I merge this? Given that the function already only returns nonnegative integers the change sounds rather cosmetic and compatible, but it is in a public header, so in pure theory somebody can rely on a specific type here.